Task #4956: Creating tag 1_17_6

+++ RTCP/Cobalt/GPUProc/src/cuda/Kernels/BeamFormerTransposeKernel.h
@@ -42,7 +42,7 @@ namespace LOFAR
       struct Parameters : Kernel::Parameters
       {
         Parameters(const Parset& ps);
-        unsigned nrTABs;
+        size_t nrTABs;
       };
 
       enum BufferType
@@ -55,11 +55,10 @@ namespace LOFAR
                              const gpu::Module &module,
                              const Buffers &buffers,
                              const Parameters &param);
-
     };
 
-    //# --------  Template specializations for KernelFactory  -------- #//
-
+    // Specialization of the KernelFactory for
+    // BeamFormerTransposeKernel
     template<> size_t
     KernelFactory<BeamFormerTransposeKernel>::bufferSize(BufferType bufferType) const;
