+++ RTCP/Cobalt/GPUProc/src/cuda/Kernels/CoherentStokesKernel.h
@@ -43,10 +43,10 @@ namespace LOFAR
       struct Parameters : Kernel::Parameters
       {
         Parameters(const Parset& ps);
-        unsigned nrTABs;
-        unsigned nrStokes;
-        unsigned timeIntegrationFactor;
-        unsigned timeParallelFactor;
+        size_t nrTABs;
+        size_t nrStokes;
+        size_t timeIntegrationFactor;
+        size_t timeParallelFactor;
       };
 
       enum BufferType
@@ -56,14 +56,13 @@ namespace LOFAR
       };
 
       CoherentStokesKernel(const gpu::Stream &stream,
-                           const gpu::Module &module,
-                           const Buffers &buffers,
-                           const Parameters &param);
-
+                             const gpu::Module &module,
+                             const Buffers &buffers,
+                             const Parameters &param);
     };
 
-    //# --------  Template specializations for KernelFactory  -------- #//
-
+    // Specialization of the KernelFactory for
+    // CoherentStokesKernel
     template<> size_t
     KernelFactory<CoherentStokesKernel>::bufferSize(BufferType bufferType) const;
